# docs/swagger-docs/swagger.json
{"swagger":"2.0","info":{"version":"1.0","title":"Reactive Notes API documentation"},"basePath":"/","tags":[{"name":"authentication-controller","description":"Authentication Controller"},{"name":"note-controller","description":"Set of endpoints for note related operations (caller must be authenticated)"},{"name":"user-controller","description":"Set of endpoints for user related operations"}],"paths":{"/login":{"post":{"tags":["authentication-controller"],"summary":"Endpoint for initial authentication","operationId":"loginUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"applicationContext.applicationName","in":"query","required":false,"type":"string"},{"name":"applicationContext.displayName","in":"query","required":false,"type":"string"},{"name":"applicationContext.id","in":"query","required":false,"type":"string"},{"name":"applicationContext.startupDate","in":"query","required":false,"type":"integer","format":"int64"},{"name":"attributes","in":"query","required":false,"type":"object"},{"name":"localeContext.locale.ISO3Country","in":"query","required":false,"type":"string"},{"name":"localeContext.locale.ISO3Language","in":"query","required":false,"type":"string"},{"name":"localeContext.locale.country","in":"query","required":false,"type":"string"},{"name":"localeContext.locale.displayCountry","in":"query","required":false,"type":"string"},{"name":"localeContext.locale.displayLanguage","in":"query","required":false,"type":"string"},{"name":"localeContext.locale.displayName","in":"query","required":false,"type":"string"},{"name":"localeContext.locale.displayScript","in":"query","required":false,"type":"string"},{"name":"localeContext.locale.displayVariant","in":"query","required":false,"type":"string"},{"name":"localeContext.locale.language","in":"query","required":false,"type":"string"},{"name":"localeContext.locale.script","in":"query","required":false,"type":"string"},{"name":"localeContext.locale.unicodeLocaleAttributes","in":"query","required":false,"type":"array","items":{"type":"string"},"collectionFormat":"multi"},{"name":"localeContext.locale.unicodeLocaleKeys","in":"query","required":false,"type":"array","items":{"type":"string"},"collectionFormat":"multi"},{"name":"localeContext.locale.variant","in":"query","required":false,"type":"string"},{"name":"logPrefix","in":"query","required":false,"type":"string"},{"name":"notModified","in":"query","required":false,"type":"boolean"},{"name":"request.id","in":"query","required":false,"type":"string"},{"name":"request.remoteAddress.address.MCGlobal","in":"query","required":false,"type":"boolean"},{"name":"request.remoteAddress.address.MCLinkLocal","in":"query","required":false,"type":"boolean"},{"name":"request.remoteAddress.address.MCNodeLocal","in":"query","required":false,"type":"boolean"},{"name":"request.remoteAddress.address.MCOrgLocal","in":"query","required":false,"type":"boolean"},{"name":"request.remoteAddress.address.MCSiteLocal","in":"query","required":false,"type":"boolean"},{"name":"request.remoteAddress.address.address","in":"query","required":false,"type":"string","format":"byte"},{"name":"request.remoteAddress.address.anyLocalAddress","in":"query","required":false,"type":"boolean"},{"name":"request.remoteAddress.address.canonicalHostName","in":"query","required":false,"type":"string"},{"name":"request.remoteAddress.address.hostAddress","in":"query","required":false,"type":"string"},{"name":"request.remoteAddress.address.hostName","in":"query","required":false,"type":"string"},{"name":"request.remoteAddress.address.linkLocalAddress","in":"query","required":false,"type":"boolean"},{"name":"request.remoteAddress.address.loopbackAddress","in":"query","required":false,"type":"boolean"},{"name":"request.remoteAddress.address.multicastAddress","in":"query","required":false,"type":"boolean"},{"name":"request.remoteAddress.address.siteLocalAddress","in":"query","required":false,"type":"boolean"},{"name":"request.remoteAddress.hostName","in":"query","required":false,"type":"string"},{"name":"request.remoteAddress.hostString","in":"query","required":false,"type":"string"},{"name":"request.remoteAddress.port","in":"query","required":false,"type":"integer","format":"int32"},{"name":"request.remoteAddress.unresolved","in":"query","required":false,"type":"boolean"},{"name":"request.sslInfo.peerCertificates[0].TBSCertificate","in":"query","required":false,"type":"string","format":"byte"},{"name":"request.sslInfo.peerCertificates[0].basicConstraints","in":"query","required":false,"type":"integer","format":"int32"},{"name":"request.sslInfo.peerCertificates[0].encoded","in":"query","required":false,"type":"string","format":"byte"},{"name":"request.sslInfo.peerCertificates[0].extendedKeyUsage","in":"query","required":false,"type":"array","items":{"type":"string"},"collectionFormat":"multi"},{"name":"request.sslInfo.peerCertificates[0].issuerDN.name","in":"query","required":false,"type":"string"},{"name":"request.sslInfo.peerCertificates[0].issuerUniqueID","in":"query","required":false,"type":"array","items":{"type":"boolean"},"collectionFormat":"multi"},{"name":"request.sslInfo.peerCertificates[0].issuerX500Principal.encoded","in":"query","required":false,"type":"string","format":"byte"},{"name":"request.sslInfo.peerCertificates[0].issuerX500Principal.name","in":"query","required":false,"type":"string"},{"name":"request.sslInfo.peerCertificates[0].keyUsage","in":"query","required":false,"type":"array","items":{"type":"boolean"},"collectionFormat":"multi"},{"name":"request.sslInfo.peerCertificates[0].notAfter","in":"query","required":false,"type":"string","format":"date-time"},{"name":"request.sslInfo.peerCertificates[0].notBefore","in":"query","required":false,"type":"string","format":"date-time"},{"name":"request.sslInfo.peerCertificates[0].serialNumber","in":"query","required":false,"type":"integer"},{"name":"request.sslInfo.peerCertificates[0].sigAlgName","in":"query","required":false,"type":"string"},{"name":"request.sslInfo.peerCertificates[0].sigAlgOID","in":"query","required":false,"type":"string"},{"name":"request.sslInfo.peerCertificates[0].sigAlgParams","in":"query","required":false,"type":"string","format":"byte"},{"name":"request.sslInfo.peerCertificates[0].signature","in":"query","required":false,"type":"string","format":"byte"},{"name":"request.sslInfo.peerCertificates[0].subjectDN.name","in":"query","required":false,"type":"string"},{"name":"request.sslInfo.peerCertificates[0].subjectUniqueID","in":"query","required":false,"type":"array","items":{"type":"boolean"},"collectionFormat":"multi"},{"name":"request.sslInfo.peerCertificates[0].subjectX500Principal.encoded","in":"query","required":false,"type":"string","format":"byte"},{"name":"request.sslInfo.peerCertificates[0].subjectX500Principal.name","in":"query","required":false,"type":"string"},{"name":"request.sslInfo.peerCertificates[0].type","in":"query","required":false,"type":"string"},{"name":"request.sslInfo.peerCertificates[0].version","in":"query","required":false,"type":"integer","format":"int32"},{"name":"request.sslInfo.sessionId","in":"query","required":false,"type":"string"},{"name":"response.statusCode","in":"query","required":false,"type":"string","enum":["100 CONTINUE","101 SWITCHING_PROTOCOLS","102 PROCESSING","103 CHECKPOINT","200 OK","201 CREATED","202 ACCEPTED","203 NON_AUTHORITATIVE_INFORMATION","204 NO_CONTENT","205 RESET_CONTENT","206 PARTIAL_CONTENT","207 MULTI_STATUS","208 ALREADY_REPORTED","226 IM_USED","300 MULTIPLE_CHOICES","301 MOVED_PERMANENTLY","302 FOUND","302 MOVED_TEMPORARILY","303 SEE_OTHER","304 NOT_MODIFIED","305 USE_PROXY","307 TEMPORARY_REDIRECT","308 PERMANENT_REDIRECT","400 BAD_REQUEST","401 UNAUTHORIZED","402 PAYMENT_REQUIRED","403 FORBIDDEN","404 NOT_FOUND","405 METHOD_NOT_ALLOWED","406 NOT_ACCEPTABLE","407 PROXY_AUTHENTICATION_REQUIRED","408 REQUEST_TIMEOUT","409 CONFLICT","410 GONE","411 LENGTH_REQUIRED","412 PRECONDITION_FAILED","413 PAYLOAD_TOO_LARGE","413 REQUEST_ENTITY_TOO_LARGE","414 URI_TOO_LONG","414 REQUEST_URI_TOO_LONG","415 UNSUPPORTED_MEDIA_TYPE","416 REQUESTED_RANGE_NOT_SATISFIABLE","417 EXPECTATION_FAILED","418 I_AM_A_TEAPOT","419 INSUFFICIENT_SPACE_ON_RESOURCE","420 METHOD_FAILURE","421 DESTINATION_LOCKED","422 UNPROCESSABLE_ENTITY","423 LOCKED","424 FAILED_DEPENDENCY","426 UPGRADE_REQUIRED","428 PRECONDITION_REQUIRED","429 TOO_MANY_REQUESTS","431 REQUEST_HEADER_FIELDS_TOO_LARGE","451 UNAVAILABLE_FOR_LEGAL_REASONS","500 INTERNAL_SERVER_ERROR","501 NOT_IMPLEMENTED","502 BAD_GATEWAY","503 SERVICE_UNAVAILABLE","504 GATEWAY_TIMEOUT","505 HTTP_VERSION_NOT_SUPPORTED","506 VARIANT_ALSO_NEGOTIATES","507 INSUFFICIENT_STORAGE","508 LOOP_DETECTED","509 BANDWIDTH_LIMIT_EXCEEDED","510 NOT_EXTENDED","511 NETWORK_AUTHENTICATION_REQUIRED"]}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/notes":{"get":{"tags":["note-controller"],"summary":"Returns all notes of the requesting user","operationId":"getAllUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/NoteDTO"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["note-controller"],"summary":"Creates and returns the given note","operationId":"saveUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"noteDTO","description":"noteDTO","required":true,"schema":{"$ref":"#/definitions/NoteDTO"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/NoteDTO"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["note-controller"],"summary":"Updates the note with the id of the given note based on the given note, updates the lastEdit field and returns the updated note","operationId":"updateUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"noteDTO","description":"noteDTO","required":true,"schema":{"$ref":"#/definitions/NoteDTO"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/NoteDTO"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/notes/export":{"get":{"tags":["note-controller"],"summary":"Returns all notes of the requesting user as JSON without any ids","operationId":"exportNotesUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/NoteDTO"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/notes/import":{"post":{"tags":["note-controller"],"summary":"Takes a list of notes as a JSON string and saves them all with the calling user as the owner","operationId":"importNotesUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"noteDTOs","description":"noteDTOs","required":true,"schema":{"type":"array","items":{"$ref":"#/definitions/NoteDTO"}}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ResponseEntity"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/notes/{id}":{"get":{"tags":["note-controller"],"summary":"Returns a single note based on the given id","operationId":"getByIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/NoteDTO"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["note-controller"],"summary":"Deletes the note with the given id","operationId":"deleteUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Mono«Void»"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/users":{"post":{"tags":["user-controller"],"summary":"Creates the given user","operationId":"saveUsingPOST_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"userDTO","description":"userDTO","required":true,"schema":{"$ref":"#/definitions/UserDTO"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ResponseEntity"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["user-controller"],"summary":"Updates the user with the id of the given user based on the given user","operationId":"updateUsingPUT_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"userDTO","description":"userDTO","required":true,"schema":{"$ref":"#/definitions/UserDTO"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserDTO"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/users/{id}":{"delete":{"tags":["user-controller"],"summary":"Deletes the user with the given id","operationId":"deleteUsingDELETE_1","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Mono«Void»"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}}},"definitions":{"Flux«NoteDTO»":{"type":"object","properties":{"prefetch":{"type":"integer","format":"int32"}},"title":"Flux«NoteDTO»"},"Mono«NoteDTO»":{"type":"object","title":"Mono«NoteDTO»"},"Mono«ResponseEntity«object»»":{"type":"object","title":"Mono«ResponseEntity«object»»"},"Mono«UserDTO»":{"type":"object","title":"Mono«UserDTO»"},"Mono«Void»":{"type":"object","title":"Mono«Void»"},"Mono«string»":{"type":"object","title":"Mono«string»"},"NoteDTO":{"type":"object","required":["color","content","createdAt","lastEdit","title"],"properties":{"color":{"type":"string","enum":["RED","ORANGE","YELLOW","GREEN","BLUE","PURPLE","GREY"]},"content":{"type":"object"},"createdAt":{"type":"string","format":"date-time"},"id":{"type":"string"},"lastEdit":{"type":"string","format":"date-time"},"title":{"type":"string"}},"title":"NoteDTO"},"ResponseEntity":{"type":"object","properties":{"body":{"type":"object"},"statusCode":{"type":"string","enum":["100 CONTINUE","101 SWITCHING_PROTOCOLS","102 PROCESSING","103 CHECKPOINT","200 OK","201 CREATED","202 ACCEPTED","203 NON_AUTHORITATIVE_INFORMATION","204 NO_CONTENT","205 RESET_CONTENT","206 PARTIAL_CONTENT","207 MULTI_STATUS","208 ALREADY_REPORTED","226 IM_USED","300 MULTIPLE_CHOICES","301 MOVED_PERMANENTLY","302 FOUND","302 MOVED_TEMPORARILY","303 SEE_OTHER","304 NOT_MODIFIED","305 USE_PROXY","307 TEMPORARY_REDIRECT","308 PERMANENT_REDIRECT","400 BAD_REQUEST","401 UNAUTHORIZED","402 PAYMENT_REQUIRED","403 FORBIDDEN","404 NOT_FOUND","405 METHOD_NOT_ALLOWED","406 NOT_ACCEPTABLE","407 PROXY_AUTHENTICATION_REQUIRED","408 REQUEST_TIMEOUT","409 CONFLICT","410 GONE","411 LENGTH_REQUIRED","412 PRECONDITION_FAILED","413 PAYLOAD_TOO_LARGE","413 REQUEST_ENTITY_TOO_LARGE","414 URI_TOO_LONG","414 REQUEST_URI_TOO_LONG","415 UNSUPPORTED_MEDIA_TYPE","416 REQUESTED_RANGE_NOT_SATISFIABLE","417 EXPECTATION_FAILED","418 I_AM_A_TEAPOT","419 INSUFFICIENT_SPACE_ON_RESOURCE","420 METHOD_FAILURE","421 DESTINATION_LOCKED","422 UNPROCESSABLE_ENTITY","423 LOCKED","424 FAILED_DEPENDENCY","426 UPGRADE_REQUIRED","428 PRECONDITION_REQUIRED","429 TOO_MANY_REQUESTS","431 REQUEST_HEADER_FIELDS_TOO_LARGE","451 UNAVAILABLE_FOR_LEGAL_REASONS","500 INTERNAL_SERVER_ERROR","501 NOT_IMPLEMENTED","502 BAD_GATEWAY","503 SERVICE_UNAVAILABLE","504 GATEWAY_TIMEOUT","505 HTTP_VERSION_NOT_SUPPORTED","506 VARIANT_ALSO_NEGOTIATES","507 INSUFFICIENT_STORAGE","508 LOOP_DETECTED","509 BANDWIDTH_LIMIT_EXCEEDED","510 NOT_EXTENDED","511 NETWORK_AUTHENTICATION_REQUIRED"]},"statusCodeValue":{"type":"integer","format":"int32"}},"title":"ResponseEntity"},"UserDTO":{"type":"object","required":["password","username"],"properties":{"id":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"roles":{"type":"array","items":{"type":"string","enum":["USER","ADMIN"]}}},"title":"UserDTO"}}}